Remove unnecssary additional platforms
The matrixed build isn't really necessary for a library testing itself. The
ubuntu builder (which is the cheapest machine type) is sufficient.

The arduino_ci project itself runs tests on all 3 platforms to ensure that
no matter which platform an end-user might have for a development machine,
it will still build properly.